feat(households): handle ID clash on create (#72)
