--- lib/create_household/create_household.dart
@@ -1,2 +1,0 @@
-export 'bloc/create_household_bloc.dart';
-export 'view/view.dart';
