--- lib/create_household/view/view.dart
@@ -1,1 +1,0 @@
-export 'create_household.dart';
