+++ lib/households/view/view.dart
@@ -1,1 +1,2 @@
 export 'households.dart';
+export 'create_household.dart';
