+++ lib/navigation/models/page_router.dart
@@ -4,7 +4,6 @@ import 'package:gibsonify/home/home.dart';
 import 'package:gibsonify/recipe/recipe.dart';
 import 'package:gibsonify/settings/settings.dart';
 import 'package:gibsonify/login/login.dart';
-import 'package:gibsonify/create_household/view/create_household.dart';
 import 'package:gibsonify/household/household.dart';
 
 class PageRouter {
@@ -16,7 +15,6 @@ class PageRouter {
   static const collectionsHelp = '/collectionshelp';
   static const recipesHelp = '/recipeshelp';
   static const settingsHelp = '/settingshelp';
-  static const createHousehold = 'createhousehold';
 
   static const householdPrefix = '/household';
   static const householdInitialRoute = '/household/$viewHousehold';
@@ -66,8 +64,6 @@ class PageRouter {
         return _buildRoute(const RecipesHelpPage());
       case settingsHelp:
         return _buildRoute(const SettingsHelpPage());
-      case createHousehold:
-        return _buildRoute(const CreateHouseholdPage());
       default:
         final route = routeSettings.name!;
         if (route.startsWith(householdPrefix)) {
